// libs/xplajs/src/ethermint/evm/v1/evm.d.ts
export interface Params {
    /**
     * evm_denom represents the token denomination used to run the EVM state
     * transitions.
     */
    evmDenom: string;
    /** enable_create toggles state transitions that use the vm.Create function */
    enableCreate: boolean;
    /** enable_call toggles state transitions that use the vm.Call function */
    enableCall: boolean;
    /** extra_eips defines the additional EIPs for the vm.Config */
    extraEips: bigint[];
    /** chain_config defines the EVM chain configuration parameters */
    chainConfig: ChainConfig;
    /**
     * allow_unprotected_txs defines if replay-protected (i.e non EIP155
     * signed) transactions can be executed on the state machine.
     */
    allowUnprotectedTxs: boolean;
}
export interface ParamsProtoMsg {
    typeUrl: "/ethermint.evm.v1.Params";
    value: Uint8Array;
}
/** Params defines the EVM module parameters */
export interface ParamsAmino {
    /**
     * evm_denom represents the token denomination used to run the EVM state
     * transitions.
     */
    evm_denom: string;
    /** enable_create toggles state transitions that use the vm.Create function */
    enable_create: boolean;
    /** enable_call toggles state transitions that use the vm.Call function */
    enable_call: boolean;
    /** extra_eips defines the additional EIPs for the vm.Config */
    extra_eips: string[];
    /** chain_config defines the EVM chain configuration parameters */
    chain_config: ChainConfigAmino;
    /**
     * allow_unprotected_txs defines if replay-protected (i.e non EIP155
     * signed) transactions can be executed on the state machine.
     */
    allow_unprotected_txs: boolean;
}
export interface ParamsAminoMsg {
    type: "ethermint/x/evm/Params";
    value: ParamsAmino;
}
/**
 * ChainConfig defines the Ethereum ChainConfig parameters using *sdk.Int values
 * instead of *big.Int.
 */
export interface ChainConfig {
    /** homestead_block switch (nil no fork, 0 = already homestead) */
    homesteadBlock: string;
    /** dao_fork_block corresponds to TheDAO hard-fork switch block (nil no fork) */
    daoForkBlock: string;
    /** dao_fork_support defines whether the nodes supports or opposes the DAO hard-fork */
    daoForkSupport: boolean;
    /**
     * eip150_block: EIP150 implements the Gas price changes
     * (https://github.com/ethereum/EIPs/issues/150) EIP150 HF block (nil no fork)
     */
    eip150Block: string;
    /** eip150_hash: EIP150 HF hash (needed for header only clients as only gas pricing changed) */
    eip150Hash: string;
    /** eip155_block: EIP155Block HF block */
    eip155Block: string;
    /** eip158_block: EIP158 HF block */
    eip158Block: string;
    /** byzantium_block: Byzantium switch block (nil no fork, 0 = already on byzantium) */
    byzantiumBlock: string;
    /** constantinople_block: Constantinople switch block (nil no fork, 0 = already activated) */
    constantinopleBlock: string;
    /** petersburg_block: Petersburg switch block (nil same as Constantinople) */
    petersburgBlock: string;
    /** istanbul_block: Istanbul switch block (nil no fork, 0 = already on istanbul) */
    istanbulBlock: string;
    /** muir_glacier_block: Eip-2384 (bomb delay) switch block (nil no fork, 0 = already activated) */
    muirGlacierBlock: string;
    /** berlin_block: Berlin switch block (nil = no fork, 0 = already on berlin) */
    berlinBlock: string;
    /** london_block: London switch block (nil = no fork, 0 = already on london) */
    londonBlock: string;
    /** arrow_glacier_block: Eip-4345 (bomb delay) switch block (nil = no fork, 0 = already activated) */
    arrowGlacierBlock: string;
    /** gray_glacier_block: EIP-5133 (bomb delay) switch block (nil = no fork, 0 = already activated) */
    grayGlacierBlock: string;
    /** merge_netsplit_block: Virtual fork after The Merge to use as a network splitter */
    mergeNetsplitBlock: string;
    /** shanghai_block switch block (nil = no fork, 0 = already on shanghai) */
    shanghaiBlock: string;
    /** cancun_block switch block (nil = no fork, 0 = already on cancun) */
    cancunBlock: string;
}
export interface ChainConfigProtoMsg {
    typeUrl: "/ethermint.evm.v1.ChainConfig";
    value: Uint8Array;
}
/**
 * ChainConfig defines the Ethereum ChainConfig parameters using *sdk.Int values
 * instead of *big.Int.
 */
export interface ChainConfigAmino {
    /** homestead_block switch (nil no fork, 0 = already homestead) */
    homestead_block: string;
    /** dao_fork_block corresponds to TheDAO hard-fork switch block (nil no fork) */
    dao_fork_block: string;
    /** dao_fork_support defines whether the nodes supports or opposes the DAO hard-fork */
    dao_fork_support: boolean;
    /**
     * eip150_block: EIP150 implements the Gas price changes
     * (https://github.com/ethereum/EIPs/issues/150) EIP150 HF block (nil no fork)
     */
    eip150_block: string;
    /** eip150_hash: EIP150 HF hash (needed for header only clients as only gas pricing changed) */
    eip150_hash: string;
    /** eip155_block: EIP155Block HF block */
    eip155_block: string;
    /** eip158_block: EIP158 HF block */
    eip158_block: string;
    /** byzantium_block: Byzantium switch block (nil no fork, 0 = already on byzantium) */
    byzantium_block: string;
    /** constantinople_block: Constantinople switch block (nil no fork, 0 = already activated) */
    constantinople_block: string;
    /** petersburg_block: Petersburg switch block (nil same as Constantinople) */
    petersburg_block: string;
    /** istanbul_block: Istanbul switch block (nil no fork, 0 = already on istanbul) */
    istanbul_block: string;
    /** muir_glacier_block: Eip-2384 (bomb delay) switch block (nil no fork, 0 = already activated) */
    muir_glacier_block: string;
    /** berlin_block: Berlin switch block (nil = no fork, 0 = already on berlin) */
    berlin_block: string;
    /** london_block: London switch block (nil = no fork, 0 = already on london) */
    london_block: string;
    /** arrow_glacier_block: Eip-4345 (bomb delay) switch block (nil = no fork, 0 = already activated) */
    arrow_glacier_block: string;
    /** gray_glacier_block: EIP-5133 (bomb delay) switch block (nil = no fork, 0 = already activated) */
    gray_glacier_block: string;
    /** merge_netsplit_block: Virtual fork after The Merge to use as a network splitter */
    merge_netsplit_block: string;
    /** shanghai_block switch block (nil = no fork, 0 = already on shanghai) */
    shanghai_block: string;
    /** cancun_block switch block (nil = no fork, 0 = already on cancun) */
    cancun_block: string;
}
export interface ChainConfigAminoMsg {
    type: "/ethermint.evm.v1.ChainConfig";
    value: ChainConfigAmino;
}
/** State represents a single Storage key value pair item. */
export interface State {
    /** key is the stored key */
    key: string;
    /** value is the stored value for the given key */
    value: string;
}
export interface StateProtoMsg {
    typeUrl: "/ethermint.evm.v1.State";
    value: Uint8Array;
}
/** State represents a single Storage key value pair item. */
export interface StateAmino {
    /** key is the stored key */
    key: string;
    /** value is the stored value for the given key */
    value: string;
}
export interface StateAminoMsg {
    type: "/ethermint.evm.v1.State";
    value: StateAmino;
}
// ...
export declare const Params: {
    typeUrl: string;
    aminoType: string;
    is(o: any): o is Params;
    isAmino(o: any): o is ParamsAmino;
    encode(message: Params, writer?: BinaryWriter): BinaryWriter;
    decode(input: BinaryReader | Uint8Array, length?: number): Params;
    fromPartial(object: DeepPartial<Params>): Params;
    fromAmino(object: ParamsAmino): Params;
    toAmino(message: Params): ParamsAmino;
    fromAminoMsg(object: ParamsAminoMsg): Params;
    toAminoMsg(message: Params): ParamsAminoMsg;
    fromProtoMsg(message: ParamsProtoMsg): Params;
    toProto(message: Params): Uint8Array;
    toProtoMsg(message: Params): ParamsProtoMsg;
};
export declare const ChainConfig: {
    typeUrl: string;
    is(o: any): o is ChainConfig;
    isAmino(o: any): o is ChainConfigAmino;
    encode(message: ChainConfig, writer?: BinaryWriter): BinaryWriter;
    decode(input: BinaryReader | Uint8Array, length?: number): ChainConfig;
    fromPartial(object: DeepPartial<ChainConfig>): ChainConfig;
    fromAmino(object: ChainConfigAmino): ChainConfig;
    toAmino(message: ChainConfig): ChainConfigAmino;
    fromAminoMsg(object: ChainConfigAminoMsg): ChainConfig;
    fromProtoMsg(message: ChainConfigProtoMsg): ChainConfig;
    toProto(message: ChainConfig): Uint8Array;
    toProtoMsg(message: ChainConfig): ChainConfigProtoMsg;
};
export declare const State: {
    typeUrl: string;
    is(o: any): o is State;
    isAmino(o: any): o is StateAmino;
    encode(message: State, writer?: BinaryWriter): BinaryWriter;
    decode(input: BinaryReader | Uint8Array, length?: number): State;
    fromPartial(object: DeepPartial<State>): State;
    fromAmino(object: StateAmino): State;
    toAmino(message: State): StateAmino;
    fromAminoMsg(object: StateAminoMsg): State;
    fromProtoMsg(message: StateProtoMsg): State;
    toProto(message: State): Uint8Array;
    toProtoMsg(message: State): StateProtoMsg;
};